Update README.md
Was a confused by the wording and spent a bit of time thinking that I was supposed to put the npSettings into /plugin/built.sbt. Added an e.g. to clarify that this should be a separate file. 

Might be just my coffee addled brain however. Ignore PR if you think unnecessary.